Lab_08/field-reader/requirements.txt
psycopg2==2.9.1
kafka-python==2.0.2